Restricting Access with Permissions (Django Rest framework)

+++ src/blog/api/views.py
@@ -1,6 +1,8 @@
 from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
+
 
 from account.models import Account
 from blog.models import BlogPost
@@ -8,6 +10,7 @@ from blog.api.serializers import BlogPostSerialize
 
 
 @api_view(['GET',])
+@permission_classes((IsAuthenticated,))
 def api_detail_blog_view(request, slug):
     try:
         blog_post = BlogPost.objects.get(slug=slug)
@@ -20,11 +23,16 @@ def api_detail_blog_view(request, slug):
 
 
 @api_view(['PUT',])
+@permission_classes((IsAuthenticated,))
 def api_update_blog_view(request, slug):
     try:
         blog_post = BlogPost.objects.get(slug=slug)
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+    user = request.user
+    if blog_post.author != user:
+        return Response({'response': "You don't have permission to edit this post"})
 
     if request.method == 'PUT':
         serializer = BlogPostSerialize(blog_post, data=request.data)
@@ -33,15 +41,20 @@ def api_update_blog_view(request, slug):
             serializer.save()
             data["success"] = "update successful"
             return Response(data=data)
-        return Response(serializer.error, status=status.HTTP_400_BAD_REQUEST)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 @api_view(['DELETE', ])
+@permission_classes((IsAuthenticated,))
 def api_delete_blog_view(request, slug):
     try:
         blog_post = BlogPost.objects.get(slug=slug)
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+    user = request.user
+    if blog_post.author != user:
+        return Response({'response': "You don't have permission to delete this post"})
 
     if request.method == 'DELETE':
         operation = blog_post.delete()
@@ -55,8 +68,9 @@ def api_delete_blog_view(request, slug):
 
 
 @api_view(['POST',])
+@permission_classes((IsAuthenticated,))
 def api_create_blog_view(request):
-    account = Account.objects.get(pk=1)
+    account = request.user
     blog_post = BlogPost(author=account)
 
     if request.method == "POST":
